Split styled components into files
