+++ src/components/common/Nav/Nav.styled.js
@@ -1,0 +1,40 @@
+import styled, { css } from 'styled-components';
+
+import mobile from 'utils/styles/mobile';
+import DropdownArrow from 'components/common/DropdownArrow';
+import Link from 'components/common/Link';
+
+export const LogoLink = styled(Link)`
+  padding-left: 1rem;
+`;
+
+export const DesktopContainer = styled.div`
+  display: ${({ authenticated }) => (authenticated ? 'block' : 'none')};
+
+  ${mobile(css`
+    display: none;
+  `)}
+`;
+
+export const StyledDropdownArrow = styled(DropdownArrow)`
+  height: 1.5rem;
+  width: 1.5rem;
+  position: absolute;
+  top: 0;
+  left: 0;
+`;
+
+export default styled.nav`
+  display: flex;
+  align-items: center;
+  background: ${({ theme }) => theme.background.primary};
+  height: 3rem;
+
+  ${mobile(css`
+    justify-content: space-between;
+    position: fixed;
+    width: 100%;
+    left: 0;
+    top: 0;
+  `)}
+`;
